Fix solutions for 1001 and 1002
The 1001 gave me more pause than expected, but that was due to not
realizing that my computer is 64 bit and the judge is 32 bit. The scanf
string was the culprit. The solution to 1002 was easier once I realized
how to do it, but it was not as easy as it first seemed.