Add beginning of optimized AoC in C++
This is the beginning of getting back into C++ and really learning the
details of it. I'll be using the newest compiler versions and I'll be
trying to get the most speed possible.

This is also a first time using please for more advanced things, so we
will see how this goes.