Speed up 2015 Day 3
This converts 2015 Day 3 to use a bitset instead of an set. This speeds
thigns up considerably, but to get it to work, I've had to make some
assumptions about the inputs. I'm not sure if I like this, but I'll keep
it as it is. The goal for now is to write the fastest solution that will
work on all aoc server outputs, without being to cheeky about it.

This change also updates the version of the fmt library to the latest
main. There has been some speedups since the last version and we want
that.