2021 Day 4
Most of the code here deals with parsing the bingo format, given that we
don't know if the numbers are padded or not. The way I solve it
logically is just to compute the turn a board will end as soon as I see
it and then just keep track of the last and first boards as we iterate
through.

This is quite fast, but I would like to see if there are some SIMD
expressions that could make the comparisons faster.